+++ scripts/macros/legacy/classFeatures/wizard/schoolOfEvocation/empoweredEvocation.mjs
@@ -1,9 +1,19 @@
-import {automationUtils, constants, workflowUtils} from '../../../../../proxy.mjs';
+import {automationUtils, constants, DamageBonus, documentUtils, workflowUtils} from '../../../../../proxy.mjs';
 async function damage({document, workflow}) {
     if (workflow.item?.type !== 'spell') return;
     if (workflow.item.system.school !== 'evo') return;
+    const multiSingleTarget = workflow.workflowOptions['chris-premades']?.multiSingleTarget;
+    if (multiSingleTarget && document.flags['chris-premades']?.lastUse === multiSingleTarget.rollID) return;
     const ability = automationUtils.getConfigValue(document, 'spellcastingAbility');
-    await workflowUtils.bonusDamage(workflow, '@abilities.' + ability + '.mod');
+    return new DamageBonus(document, {
+        formula: '@abilities.' + ability + '.mod',
+        optional: multiSingleTarget ? multiSingleTarget.remainingAttacks > 1 : false
+    })
+        .withOnUse(async () => {
+            if (multiSingleTarget) await documentUtils.setFlag(document, 'chris-premades', 'lastUse', multiSingleTarget.rollID);
+            await workflowUtils.completeItemUse(document);
+        })
+        .initialize();
 }
 export const empoweredEvocation = {
     name: 'Empowered Evocation',
@@ -11,7 +21,7 @@ export const empoweredEvocation = {
     rules: '2014',
     roll: [
         {
-            pass: 'actorDamageRollComplete',
+            pass: 'actorOptionalBonusDamage',
             macro: damage,
             priority: 50
         }